requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
pytest==6.2.5
python-jose==3.3.0
python-dotenv==0.19.0
werkzeug==2.3.7
requests==2.31.0
pymysql==1.1.0
PyJWT==2.8.0 